+++ lm_engine/hf_adapter.py
@@ -4,18 +4,40 @@
 
 from __future__ import annotations
 
+import os
+
 import torch
 from transformers import GenerationMixin, PretrainedConfig, PreTrainedModel
 from transformers.modeling_outputs import CausalLMOutputWithPast as _HFCausalLMOutputWithPast
 
+from .arguments import LoadArgs, MixedPrecisionArgs, UnshardingArgs
 from .generation_cache import GenerationCache
 from .loss import get_autoregressive_language_modeling_loss
 from .mixins.dense.main import CausalLMModelMixin
 from .model_config import CommonConfig
 from .modeling_utils import ParameterizedEmbedding, ParameterizedLinear
+from .parallel import ProcessGroupManager
+from .parameter import (
+    _INIT_MARKER,
+    get_named_parameters_and_buffers,
+    get_parameter_marker_maps,
+    is_parameter_initialized,
+    set_parameter_marker_maps,
+)
+from .utils import SafeTensorsWeightsManager, torch_dtype_to_string
 
 
 _MODEL_TYPE_TO_CAUSAL_LM_CLASS: dict[str, type[CausalLMModelMixin]] = {}
+_HF_USELESS_STUFF = [
+    "revision",
+    "subfolder",
+    "gguf_file",
+    "quantization_config",
+    "max_memory",
+    "name_or_path",
+    "trust_remote_code",
+    "adapter_kwargs",
+]
 
 
 class LLMAdapter_HFConfig(PretrainedConfig):
@@ -36,15 +58,16 @@ class LLMAdapter_HFConfig(PretrainedConfig):
 class LLMAdapter_HF(PreTrainedModel, GenerationMixin):
     """Adapts an already-built lm_engine `CausalLMModelMixin` model (e.g. `GPTBaseForCausalLM`) to the
     `transformers.PreTrainedModel` + `GenerationMixin` interface, so it can be dropped into `model.generate(...)`,
-    `transformers.pipeline(...)`, or any HuggingFace-based evaluation harness without modification.
+    `transformers.pipeline(...)`, or any HuggingFace-based evaluation harness without modification. It also owns
+    checkpoint I/O (`from_pretrained`/`save_pretrained`) for all custom lm_engine architectures; the lm_engine
+    training loop bypasses this class and constructs the raw model directly (see `register_hf.get_causal_lm_class`).
 
-    This is an inference-only adapter. It does not support training (no pipeline/tensor/sequence parallelism,
-    no padding-free packed sequences, no loss-scaled aux loss handling): use the wrapped lm_engine model, its own
-    `mixins.dense.main.CausalLMModelMixin.generate`, and the lm_engine training loop directly for anything else.
+    This is an inference-only adapter: `forward()`/`generate()` assume a single-stage, non-padding-free model; use
+    the lm_engine training loop directly for anything else (loss-scaled aux loss, `PipelineParallelOutput`, packed
+    padding-free sequences).
 
     Example:
-        model = GPTBaseForCausalLM.from_pretrained(path)
-        hf_model = LLMAdapter_HF(model)
+        hf_model = LLMAdapter_HF.from_pretrained(path, config=config)
         hf_model.generate(input_ids=input_ids, attention_mask=attention_mask, max_new_tokens=20)
     """
 
@@ -52,12 +75,10 @@ class LLMAdapter_HF(PreTrainedModel, GenerationMixin):
     base_model_prefix = "model"
 
     def __init__(self, model: CausalLMModelMixin) -> LLMAdapter_HF:
-        # NOTE: pipeline parallelism is allowed through for now so `model_wrapper/base.py::get_loss` can wrap a
-        # pipeline-parallel model just to reach `get_output_embeddings()` (only ever called on the last stage,
-        # where `lm_head`/`wte` are guaranteed to exist). `forward()`/`generate()` still assume a single-stage
-        # model and were never updated to handle `PipelineParallelOutput` — don't call them through this adapter
-        # on a pipeline-parallel model.
-        assert not model.use_padding_free_transformer, "LLMAdapter_HF does not support the padding-free transformer"
+        # NOTE: pipeline-parallel and padding-free-transformer models are allowed through for now so callers
+        # (e.g. `model_wrapper/base.py::get_loss`) can wrap one just to reach `get_output_embeddings()` etc.
+        # `forward()`/`generate()` still assume a single-stage, non-padding-free model (no `PipelineParallelOutput`
+        # handling, no `cu_seqlens`/`max_seqlen` params) — don't call them through this adapter on such a model.
 
         super().__init__(type(self).config_class.from_common_config(model.config))
 
@@ -81,14 +102,80 @@ class LLMAdapter_HF(PreTrainedModel, GenerationMixin):
 
     @classmethod
     def from_pretrained(
-        cls, pretrained_model_name_or_path: str, *model_args, config: CommonConfig | None = None, **kwargs
+        cls,
+        pretrained_model_name_or_path: str,
+        *model_args,
+        config: CommonConfig | None = None,
+        dtype: torch.dtype = torch.float32,
+        iteration: int | None = None,
+        **kwargs,
     ) -> LLMAdapter_HF:
+        assert len(model_args) == 0
         assert config is not None, "LLMAdapter_HF.from_pretrained needs `config` (an lm_engine CommonConfig)"
 
-        model = cls._get_causal_lm_class(config).from_pretrained(
-            pretrained_model_name_or_path, *model_args, config=config, **kwargs
-        )
+        model_class = cls._get_causal_lm_class(config)
+
+        # drop useless stuff
+        for k in _HF_USELESS_STUFF:
+            kwargs.pop(k, None)
+
+        num_pipeline_stages = kwargs.pop("num_pipeline_stages", 1)
+        pipeline_stage_id = kwargs.pop("pipeline_stage_id", 0)
+
+        if os.path.isfile(os.path.join(pretrained_model_name_or_path, "latest_checkpointed_iteration.json")):
+            # lazy import avoids circular dependency (checkpointing → model_wrapper → hf_adapter)
+            from .checkpointing import load_checkpoint_and_unshard
+
+            assert not ProcessGroupManager.is_tensor_parallel_enabled()
+
+            unshard_args = UnshardingArgs(
+                load_args=LoadArgs(load_path=pretrained_model_name_or_path, iteration=iteration),
+                mixed_precision_args=MixedPrecisionArgs(dtype=torch_dtype_to_string(dtype)),
+                unsharded_path="",
+            )
+
+            model_wrapper, _, _ = load_checkpoint_and_unshard(unshard_args)
+            model = model_wrapper.model
+        else:
+            assert iteration is None, "iteration should be None when loading from an unsharded checkpoint"
+
+            if ProcessGroupManager.is_tensor_parallel_enabled() or num_pipeline_stages > 1:
+                with torch.device("meta"):
+                    model = model_class(
+                        config, num_pipeline_stages=num_pipeline_stages, pipeline_stage_id=pipeline_stage_id, **kwargs
+                    )
+
+                for module in model.modules():
+                    if hasattr(module, "reset_parameters"):
+                        module.reset_parameters()
+
+                marker_maps = get_parameter_marker_maps([model], extra_markers=[_INIT_MARKER])
+
+                model = model.to(dtype=dtype)
+                model = model.to_empty(device=torch.cuda.current_device())
+                model.load_from_safetensors_weights_manager(SafeTensorsWeightsManager(pretrained_model_name_or_path))
+            else:
+                model = model_class(config, **kwargs)
+                marker_maps = get_parameter_marker_maps([model], extra_markers=[_INIT_MARKER])
+
+                model = model.to(dtype=dtype)
+                model.load_state_dict(SafeTensorsWeightsManager(pretrained_model_name_or_path).state_dict())
+
+        device_map = kwargs.pop("device_map", {"": None})
+        assert len(device_map) == 1
+        assert len(kwargs) == 0
+
+        model = model.to(device_map[""])
+        set_parameter_marker_maps([model], marker_maps)
+
+        for param_name, param in get_named_parameters_and_buffers(model):
+            assert is_parameter_initialized(param), f"{param_name} is not initialized"
+
         return cls(model)
+
+    def save_pretrained(self, save_directory: str, **kwargs) -> None:
+        self.model.config.save_pretrained(save_directory)
+        SafeTensorsWeightsManager.save_state_dict(self.model.state_dict(), save_directory)
 
     @classmethod
     def _supports_default_dynamic_cache(cls) -> bool:
